Move copied code to separate file
