+++ crates/runtime/src/lib.rs
@@ -1,23 +1,19 @@
+use crate::vm::{MemBuffer, cell_ref_to_relocatable, extract_relocatable, vm_get_range};
 use anyhow::Result;
 use blockifier::execution::deprecated_syscalls::DeprecatedSyscallSelector;
 use blockifier::execution::syscalls::hint_processor::SyscallHintProcessor;
 use blockifier::state::errors::StateError;
 use cairo_lang_casm::hints::{Hint, StarknetHint};
-use cairo_lang_casm::operand::{
-    BinOpOperand, CellRef, DerefOrImmediate, Operation, Register, ResOperand,
-};
+use cairo_lang_casm::operand::{CellRef, ResOperand};
 use cairo_lang_utils::bigint::BigIntAsHex;
-use cairo_lang_utils::extract_matches;
-use cairo_vm::Felt252;
 use cairo_vm::hint_processor::hint_processor_definition::{
     HintProcessor, HintProcessorLogic, HintReference,
 };
 use cairo_vm::serde::deserialize_program::ApTracking;
 use cairo_vm::types::exec_scope::ExecutionScopes;
-use cairo_vm::types::relocatable::{MaybeRelocatable, Relocatable};
+use cairo_vm::types::relocatable::Relocatable;
 use cairo_vm::vm::errors::hint_errors::HintError;
 use cairo_vm::vm::errors::hint_errors::HintError::CustomHint;
-use cairo_vm::vm::errors::memory_errors::MemoryError;
 use cairo_vm::vm::errors::vm_errors::VirtualMachineError;
 use cairo_vm::vm::runners::cairo_runner::{ResourceTracker, RunResources};
 use cairo_vm::vm::vm_core::VirtualMachine;
@@ -35,6 +31,7 @@ use std::io;
 use thiserror::Error;
 
 pub mod starknet;
+mod vm;
 
 // from core/src/starknet/testing.cairo
 const CAIRO_TEST_CHEATCODES: [&str; 14] = [
@@ -113,133 +110,6 @@ fn parse_selector(selector: &BigIntAsHex) -> Result<String, HintError> {
     let selector = std::str::from_utf8(selector)
         .map_err(|_| CustomHint(Box::from("Failed to parse the cheatcode selector")))?;
     Ok(String::from(selector))
-}
-
-// TODO copied code
-fn cell_ref_to_relocatable(cell_ref: &CellRef, vm: &VirtualMachine) -> Relocatable {
-    let base = match cell_ref.register {
-        Register::AP => vm.get_ap(),
-        Register::FP => vm.get_fp(),
-    };
-    (base + (cell_ref.offset as i32)).unwrap()
-}
-
-// TODO copied code
-/// Extracts a parameter assumed to be a buffer.
-fn extract_buffer(buffer: &ResOperand) -> (&CellRef, Felt252) {
-    let (cell, base_offset) = match buffer {
-        ResOperand::Deref(cell) => (cell, 0.into()),
-        ResOperand::BinOp(BinOpOperand {
-            op: Operation::Add,
-            a,
-            b,
-        }) => (
-            a,
-            extract_matches!(b, DerefOrImmediate::Immediate)
-                .clone()
-                .value
-                .into(),
-        ),
-        _ => panic!("Illegal argument for a buffer."),
-    };
-    (cell, base_offset)
-}
-
-// TODO copied code
-/// Fetches the value of a cell plus an offset from the vm, useful for pointers.
-fn get_ptr(
-    vm: &VirtualMachine,
-    cell: &CellRef,
-    offset: &Felt252,
-) -> std::result::Result<Relocatable, VirtualMachineError> {
-    Ok((vm.get_relocatable(cell_ref_to_relocatable(cell, vm))? + offset)?)
-}
-
-// TODO copied code
-/// Extracts a parameter assumed to be a buffer, and converts it into a relocatable.
-fn extract_relocatable(
-    vm: &VirtualMachine,
-    buffer: &ResOperand,
-) -> std::result::Result<Relocatable, VirtualMachineError> {
-    let (base, offset) = extract_buffer(buffer);
-    get_ptr(vm, base, &offset)
-}
-
-// TODO copied code
-fn vm_get_range(
-    vm: &mut VirtualMachine,
-    mut calldata_start_ptr: Relocatable,
-    calldata_end_ptr: Relocatable,
-) -> std::result::Result<Vec<Felt252>, HintError> {
-    let mut values = vec![];
-    while calldata_start_ptr != calldata_end_ptr {
-        let val = *vm.get_integer(calldata_start_ptr)?;
-        values.push(val);
-        calldata_start_ptr.offset += 1;
-    }
-    Ok(values)
-}
-
-// TODO copied code
-/// Wrapper trait for a VM owner.
-pub trait VMWrapper {
-    fn vm(&mut self) -> &mut VirtualMachine;
-}
-impl VMWrapper for VirtualMachine {
-    fn vm(&mut self) -> &mut VirtualMachine {
-        self
-    }
-}
-
-// TODO copied code
-/// A helper struct to continuously write and read from a buffer in the VM memory.
-pub struct MemBuffer<'a> {
-    /// The VM to write to.
-    /// This is a trait so that we would borrow the actual VM only once.
-    vm: &'a mut dyn VMWrapper,
-    /// The current location of the buffer.
-    pub ptr: Relocatable,
-}
-
-// TODO copied code
-impl<'a> MemBuffer<'a> {
-    /// Creates a new buffer.
-    pub fn new(vm: &'a mut dyn VMWrapper, ptr: Relocatable) -> Self {
-        Self { vm, ptr }
-    }
-
-    /// Creates a new segment and returns a buffer wrapping it.
-    pub fn new_segment(vm: &'a mut dyn VMWrapper) -> Self {
-        let ptr = vm.vm().add_memory_segment();
-        Self::new(vm, ptr)
-    }
-
-    /// Returns the current position of the buffer and advances it by one.
-    fn next(&mut self) -> Relocatable {
-        let ptr = self.ptr;
-        self.ptr += 1;
-        ptr
-    }
-
-    /// Writes a value to the current position of the buffer and advances it by one.
-    pub fn write<T: Into<MaybeRelocatable>>(
-        &mut self,
-        value: T,
-    ) -> std::result::Result<(), MemoryError> {
-        let ptr = self.next();
-        self.vm.vm().insert_value(ptr, value)
-    }
-    /// Writes an iterator of values starting from the current position of the buffer and advances
-    /// it to after the end of the written value.
-    pub fn write_data<T: Into<MaybeRelocatable>, Data: Iterator<Item = T>>(
-        &mut self,
-        data: Data,
-    ) -> std::result::Result<(), MemoryError> {
-        for value in data {
-            self.write(value)?;
-        }
-        Ok(())
-    }
 }
 
 fn fetch_cheatcode_input(
